+++ RELEASES.md
@@ -7,9 +7,10 @@ notice are preserved. This file is offered as-is, without any warranty. -->
 Version 1.15.0 (unreleased)
 ===========================
 
-  * Now [`Integer`][int-1-15] implements [`Shl`], [`Shr`], [`ShlAssign`] and
-    [`ShrAssign`] with the right-hand side types [`isize`] and [`usize`] as well
-    as the previously supported [`i32`] and [`u32`].
+  * Now [`Integer`][int-1-15] and [`Rational`][rat-1-15] implement [`Shl`],
+    [`Shr`], [`ShlAssign`] and [`ShrAssign`] with the right-hand side types
+    [`isize`] and [`usize`] as well as the previously supported [`i32`] and
+    [`u32`].
   * The new experimental feature [`num-traits`][feat-nt-1-15] was added to
     implement some traits from the [*num-traits* crate] and the [*num-integer*
     crate] ([issue 30]).
@@ -17,6 +18,7 @@ Version 1.15.0 (unreleased)
 [feat-nt-1-15]: https://tspiteri.gitlab.io/rug/dev/rug/index.html#experimental-optional-features
 [int-1-15]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Integer.html
 [issue 30]: https://gitlab.com/tspiteri/rug/-/issues/30
+[rat-1-15]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Rational.html
 
 Version 1.14.1 (2022-01-23)
 ===========================